Separate Info Component into Country and Weather components
